requirements.txt
pandas>=1.5.3
numpy>=1.24.2
notebook>=6.5.3
python-dotenv>=1.0.0
nmrglue
